Add "batchShowNodes" command line argument to benchmarks.
If enabled, the batch target nodes and count of keys directed at each node are printed once at the start of benchmarks.
This is useful to determine how increasing the cluster size also increases the number of batch target nodes, which may
result in a performance decrease for batch commands.