CLIENT-2554 Use the standard Base64 class instead of the old javax.xml.bind.DatatypeConverter for base64 encoding/decoding.  DatatypeConverter was used when bouncycastle crypto was enabled.  DatatypeConverter has been removed from newer versions of the java JDK.
Change PartitionParser to not send the trailing newline to the Base64 decoder, because this decoder does not allow the newline.

Add manifest filters to examples/benchmarks “pom.xml” to fix jar-with-dependencies jar signing errors when running with bouncycastle crypto and newer versions of the java JDK.

Fix the set_crypto sed script when running on Linux.